requirements.txt
# Core
streamlit>=1.35.0
pandas>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dotenv>=1.0.0

# Async & HTTP
httpx>=0.27.0
aiohttp>=3.9.0

# Playwright for JS-rendered pages
playwright>=1.44.0

# Phone/email validation
phonenumbers>=8.13.0
email-validator>=2.1.0

# Google Sheets (optional)
gspread>=6.0.0
google-auth>=2.29.0

# Rate limiting & retries
tenacity>=8.3.0
fake-useragent>=1.5.0

# Data
openpyxl>=3.1.0
